material_forge_env/pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-material_forge_env"
version = "0.1.0"
description = "Material Forge Env environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    # Core OpenEnv runtime (provides FastAPI server + HTTP client types)
    "openenv-core[core]>=0.2.2",
    # aiofiles is required by FastAPI's StaticFiles for async file serving
    "aiofiles>=23.0.0",
    # Although we use a custom UI, openenv-core imports gradio internally
    "gradio>=4.0.0",
    # Agent Test dependencies
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
# or: python -m material_forge_env.server.app
server = "material_forge_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["material_forge_env", "material_forge_env.server", "material_forge_env.environment", "material_forge_env.scenarios"]
package-dir = { "material_forge_env" = ".", "material_forge_env.server" = "server", "material_forge_env.environment" = "environment", "material_forge_env.scenarios" = "scenarios" }

[tool.setuptools.package-data]
"material_forge_env.server" = ["static/*", "static/**/*"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
